feat(plugin): `--node-selector` option with `install generate`
The `--node-selector` flag allows you to specify a set of node labels that the
target nodes must have in order to schedule the operator's Deployment object
onto them. By default, this is empty, meaning that the Deployment can be
scheduled on any available node. When you set this flag, Kubernetes will only
consider nodes that match the specified labels for deploying the operator.

This is useful when you want to control where the operator runs, such as
ensuring it runs on nodes with specific hardware, in a particular zone, or with
certain characteristics.

Closes #5270

Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>